requirements.txt
python-telegram-bot==21.2
requests
Pillow
qrcode